+++ crates/telio-relay/src/derp/http.rs
@@ -7,14 +7,14 @@ use super::proto::{
 use httparse::Status;
 use std::{
     convert::TryFrom,
-    fs::File,
-    io::{BufReader, Cursor, Error as IoError, ErrorKind},
+    io::{Cursor, Error as IoError, ErrorKind},
     net::SocketAddr,
     sync::Arc,
-    time::{Duration, SystemTime},
+    time::Duration,
 };
 use telio_sockets::{SocketBufSizes, SocketPool, TcpParams};
 use telio_task::io::Chan;
+use webpki_roots::TLS_SERVER_ROOTS;
 
 use crate::{Config, DerpKeepaliveConfig};
 
@@ -26,32 +26,12 @@ use tokio::{
     time::timeout,
 };
 use tokio_rustls::{
-    rustls::{
-        client::{ServerCertVerified, ServerCertVerifier, ServerName},
-        Certificate, ClientConfig, Error as TLSError, OwnedTrustAnchor, RootCertStore,
-    },
+    rustls::{client::ServerName, ClientConfig, OwnedTrustAnchor, RootCertStore},
     TlsConnector,
 };
 use url::{Host, Url};
-use webpki_roots::TLS_SERVER_ROOTS;
-
-pub(crate) struct NoVerifier;
 
 const SOCK_BUF_SZ: usize = 212992;
-
-impl ServerCertVerifier for NoVerifier {
-    fn verify_server_cert(
-        &self,
-        _end_entity: &Certificate,
-        _intermediates: &[Certificate],
-        _server_name: &ServerName,
-        _scts: &mut dyn Iterator<Item = &[u8]>,
-        _ocsp_response: &[u8],
-        _now: SystemTime,
-    ) -> Result<ServerCertVerified, TLSError> {
-        Ok(ServerCertVerified::assertion())
-    }
-}
 
 /// Max TCP packet size is 65535
 const MAX_TCP_PACKET_SIZE: usize = u16::MAX as usize;
@@ -145,38 +125,29 @@ pub async fn connect_http_and_start(
             .await
         }
         _ => {
-            let trust_anchors = TLS_SERVER_ROOTS.iter().map(|trust_anchor| {
-                OwnedTrustAnchor::from_subject_spki_name_constraints(
-                    trust_anchor.subject,
-                    trust_anchor.spki,
-                    trust_anchor.name_constraints,
-                )
-            });
+            let config = if derp_config.use_built_in_root_certificates {
+                let trust_anchors = TLS_SERVER_ROOTS.iter().map(|trust_anchor| {
+                    OwnedTrustAnchor::from_subject_spki_name_constraints(
+                        trust_anchor.subject,
+                        trust_anchor.spki,
+                        trust_anchor.name_constraints,
+                    )
+                });
+                let mut root_store = RootCertStore::empty();
+                root_store.add_trust_anchors(trust_anchors);
 
-            let mut root_store = RootCertStore::empty();
-            root_store.add_trust_anchors(trust_anchors);
+                let config = ClientConfig::builder()
+                    .with_safe_defaults()
+                    .with_root_certificates(root_store)
+                    .with_no_client_auth();
 
-            if let Some(path) = &derp_config.ca_pem_path {
-                let root_cert_file = File::open(path)?;
-                let mut root_cert_file = BufReader::new(root_cert_file);
-                let der_certs = rustls_pemfile::certs(&mut root_cert_file).unwrap_or_default();
-                root_store.add_parsable_certificates(&der_certs);
-            }
-
-            let mut config = ClientConfig::builder()
-                .with_safe_defaults()
-                .with_root_certificates(root_store)
-                .with_no_client_auth();
-
-            if derp_config.ca_pem_path.is_some() {
-                config
-                    .dangerous()
-                    .set_certificate_verifier(Arc::new(NoVerifier));
-            }
+                TlsConnector::from(Arc::new(config))
+            } else {
+                TlsConnector::from(Arc::new(rustls_platform_verifier::tls_config()))
+            };
 
             let server_name =
                 ServerName::try_from(hostname.as_str()).map_err(|_| "Invalid Server Name")?;
-            let config = TlsConnector::from(Arc::new(config));
 
             connect_and_start(
                 config.connect(server_name, stream).await?,