+++ nat-lab/tests/test_network_switch.py
@@ -162,10 +162,7 @@ async def test_mesh_network_switch(
                 adapter_type=telio.AdapterType.WindowsNativeWg,
                 is_meshnet=False,
             ),
-            marks=[
-                pytest.mark.windows,
-                pytest.mark.xfail(reason="Test is flaky - LLT-4391"),
-            ],
+            marks=[pytest.mark.windows],
         ),
         pytest.param(
             SetupParameters(
@@ -173,10 +170,7 @@ async def test_mesh_network_switch(
                 adapter_type=telio.AdapterType.WireguardGo,
                 is_meshnet=False,
             ),
-            marks=[
-                pytest.mark.windows,
-                pytest.mark.xfail(reason="Test is flaky - LLT-4391"),
-            ],
+            marks=[pytest.mark.windows],
         ),
         pytest.param(
             SetupParameters(
@@ -186,7 +180,7 @@ async def test_mesh_network_switch(
             ),
             marks=[
                 pytest.mark.mac,
-                pytest.mark.xfail(reason="Test is flaky - LLT-4685"),
+                pytest.mark.xfail(reason="Test is flaky - LLT-4696"),
             ],
         ),
     ],